# docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "version": "1.0.0",
    "title": "Water Tracker",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "description": "This is a documentation of water tracker"
  },
  "servers": [
    {
      "url": "https://water-tracker-backend-guwj.onrender.com",
      "description": "Production server"
    },
    {
      "url": "http://127.0.0.1:3000",
      "description": "Development server"
    }
  ],
...
  "paths": {
...
    "/auth/update-user": {
      "patch": {
        "tags": [
          "Auth"
        ],
        "summary": "Update user",
        "operationId": "updateUser",
        "description": "Update and add user data. This endpoint allows the authenticated user to update their profile information such as email, avatar, name, gender, weight, activity time, and desired water volume. The user must provide a valid access token in the `Authorization` header.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [],
                "properties": {
                  "email": {
                    "description": "The email address of the user.",
                    "type": "string",
                    "format": "email",
                    "example": "peter12@example.com"
                  },
                  "avatarURL": {
                    "description": "The link to the user's photo. This can be a binary file.",
                    "type": "string",
                    "format": "binary"
                  },
                  "name": {
                    "description": "The name of the user. Must be between 3 and 20 characters.",
                    "type": "string",
                    "minLength": 3,
                    "maxLength": 20,
                    "example": "Peter Pavlov"
                  },
                  "gender": {
                    "description": "The gender of the user.",
                    "type": "string",
                    "enum": [
                      "female",
                      "male"
                    ]
                  },
                  "weight": {
                    "description": "The weight of the user.",
                    "type": "number",
                    "minimum": 0,
                    "example": 60
                  },
                  "activityTime": {
                    "description": "The active time of the user, in hours per day.",
                    "type": "number",
                    "minimum": 0,
                    "example": 2
                  },
                  "desiredVolume": {
                    "description": "The new desired daily water intake of the user in liters. Must be between 0.05 and 5 liters.",
                    "type": "number",
                    "minimum": 0.05,
                    "maximum": 5,
                    "example": 2.5
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully updated the user.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "User updated successfully"
                    },
                    "data": {
                      "type": "object",
                      "required": [
                        "user"
                      ],
                      "properties": {
                        "user": {
                          "type": "object",
                          "$ref": "#/components/schemas/user"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
    "/auth/google-login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Login via Google OAuth",
        "operationId": "googleLogin",
        "description": "Authenticate a user using Google OAuth. This endpoint verifies the token received from Google and logs the user in, issuing a new access token for subsequent authenticated requests.",
        "security": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "code"
                ],
                "properties": {
                  "code": {
                    "description": "The authorization code received from Google after user login.",
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiI2NzgyY2I4ODQ4MDlhNjFhYjk2YzQyOGYiLCJlbWFpbCI6ImdvY2hha2luZHJAdWtyLm5ldCIsImlhdCI6MTczNjY0MzQyMiwiZXhwIjoxNzM2NjQzNzIyfQ.1qKbcPj7mv6vNp0DzO_qSzRBFD1Boa5nBIKkopjAQMc"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully logged in via Google OAuth and issued a new access token.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully logged in via Google OAuth!"
                    },
                    "data": {
                      "type": "object",
                      "required": [
                        "accessToken",
                        "user"
                      ],
                      "properties": {
                        "accessToken": {
                          "description": "The access token for the authenticated user.",
                          "type": "string",
                          "example": "NOfB2S2e0OVHjvJb41gxPf4Rc6gwZmblEgiC4axh"
                        },
                        "user": {
                          "type": "object",
                          "$ref": "#/components/schemas/user"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
...
    "/auth/reset-pwd": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Reset password",
        "operationId": "resetPassword",
        "description": "Reset the user's password. This endpoint allows a user to reset their password using a provided token. The user must provide a valid token and a new password. Upon successful reset, the user's password is updated. **The token used to authorize the password reset is typically received via email after requesting a password reset. Use the `/auth/send-reset-email` endpoint to request the token.",
        "security": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "token",
                  "password"
                ],
                "properties": {
                  "token": {
                    "description": "The token used to authorize the password reset. This token is typically received via email.",
                    "type": "string",
                    "example": "NOfB2S2e0OVHjvJb41gxPf4Rc6gwZmblEgiC4axh"
                  },
                  "password": {
                    "description": "The new password for the user. Must be between 8 and 50 characters.",
                    "minLength": 8,
                    "maxLength": 50,
                    "type": "string",
                    "example": "Asdf1234"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully reset the user's password.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Password was successfully reset!"
                    },
                    "data": {
                      "type": "object",
                      "example": {}
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "responses": {
      "400": {
        "description": "Bad Request Error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message",
                "data"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 400,
                  "description": "The HTTP status code for a bad request."
                },
                "message": {
                  "type": "string",
                  "example": "Bad Request Error",
                  "description": "A brief description of the error."
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "The request could not be understood by the server due to malformed syntax",
                      "description": "Detailed information about the error, specifying that the request was not understood due to incorrect syntax."
                    }
                  }
                }
              }
            }
          }
        }
      },
      "401": {
        "description": "Unauthorized error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message",
                "data"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 401,
                  "description": "The HTTP status code for an unauthorized request."
                },
                "message": {
                  "type": "string",
                  "example": "Unauthorized error",
                  "description": "A brief description of the error."
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Access token not provided",
                      "description": "Detailed information about the error, specifying that the access token was not provided."
                    }
                  }
                }
              }
            }
          }
        }
      },
      "404": {
        "description": "Not Found Error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message",
                "data"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 404,
                  "description": "The HTTP status code indicating that the requested resource was not found."
                },
                "message": {
                  "type": "string",
                  "example": "Not Found Error",
                  "description": "A brief description of the error."
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "The specified resource was not found!",
                      "description": "Detailed information about the error, specifying that the resource requested could not be found."
                    }
                  }
                }
              }
            }
          }
        }
      },
...
      "500": {
        "description": "Internal server error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message",
                "data"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 500,
                  "description": "The HTTP status code indicating that the server encountered an unexpected condition that prevented it from fulfilling the request."
                },
                "message": {
                  "type": "string",
                  "example": "Internal server error",
                  "description": "A brief description of the error."
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Internal server error",
                      "description": "Detailed information about the error, specifying that the server encountered an unexpected condition."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "schemas": {
      "user": {
        "type": "object",
        "required": [
          "_id",
          "name",
          "email",
          "avatarURL",
          "gender",
          "weight",
          "activityTime",
          "desiredVolume",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "_id": {
            "description": "The unique identifier of the user.",
            "type": "string",
            "example": "65e4decdd286b30065d54af9"
          },
          "name": {
            "description": "The name of the user.",
            "type": "string",
            "example": "Peter Pavlov"
          },
          "email": {
            "description": "The email address of the user.",
            "type": "string",
            "example": "peter12@example.com"
          },
          "avatarURL": {
            "description": "The URL link to the user's avatar photo. This can be a binary file.",
            "type": "string",
            "format": "binary"
          },
          "gender": {
            "description": "The gender of the user.",
            "type": "string",
            "enum": [
              "female",
              "male"
            ]
          },
          "weight": {
            "description": "The weight of the user in kilograms.",
            "type": "number",
            "example": 60
          },
          "activityTime": {
            "description": "The daily activity time of the user in hours.",
            "type": "number",
            "example": 2
          },
          "desiredVolume": {
            "description": "The desired daily water intake volume of the user in liters. Must be between 0.05l and 5l.",
            "type": "number",
            "example": 1.5,
            "minimum": 0.05,
            "maximum": 5
          },
          "createdAt": {
            "description": "The date and time when the user was created. Must be in ISO 8601 format.",
            "type": "string",
            "example": "2025-01-11T19:54:04.483Z"
          },
          "updatedAt": {
            "description": "The date and time when the user was last updated. Must be in ISO 8601 format.",
            "type": "string",
            "example": "2025-01-11T19:54:04.483Z"
          }
        }
      },
...
    }
  }
}